Css home and fix GetDate
Create notification.html to display notification Fix authorization and authentication dispatcher to notification when access denied. Css home.jsp and finally, remove the getSQLDatesBetween function and replace it with getWeekContainingDate, this function will get the week containing the selected date

+++ Assignment/src/java/controller/ListStudentAttendanceController.java
@@ -43,7 +43,7 @@ public class ListStudentAttendanceController extends AuthorizationController {
             StudentDBContext stuDB = new StudentDBContext();
             ArrayList<Student> students = stuDB.getStudentByGnameAndSubname(gname, subname);
 
-            List<Double> absentPercentages = new ArrayList<>();
+            List<Float> absentPercentages = new ArrayList<>();
 
             int totalSessions = dates.size();
             for (Student student : students) {
@@ -60,7 +60,7 @@ public class ListStudentAttendanceController extends AuthorizationController {
                 }
 
                 // Tính phần trăm absent và thêm vào danh sách
-                double absentPercentage = (double) totalAbsent / totalSessions * 100;
+                float absentPercentage = (float) totalAbsent / totalSessions * 100;
                 absentPercentages.add(absentPercentage);
             }
             req.setAttribute("absentPercentages", absentPercentages);
